activated derivative b test
